Cleaned up examples and benchmarks.  Found strange bug.  Tweaked Process.hs for strict bytestrings.
The bug is that DemoTransport.hs will hang when trying to "killThread"
(Pipes backend, demo3 and demo4), but ONLY under GHC 7.0.4 and 7.2.1.
Under GHC 6.12.3 and GHC 7.4.1 it works fine!

At first I thought this may be an issue with non-allocating threads not
being preempted by the runtime system (and therefore not servicing the
ThreadKilled asynchronous exception).  But it's hard to explain that
pattern of outcomes on different GHC versions.