Support self-connections in the TCP transport
This case must be treated specially, because local endpoints have a one-to-one
mapping from sockets that they are listening on to remote endpoints. For a
self-connection, the remote endpoint == the local endpoint and both sockets are
on the same local endpoint, which breaks our representation (and yields all
kinds of nastiness). A benefit of treating this case special is of course
performance.