Don't close connection on asynchronous exception
In modifyRemoteState we put the remote state in closed state when any exception
occured. This is incorrect, however. If a thread that is trying to send gets
killed, that doesn't necessary mean that the connection is now broken. Now we
close the remote endpoint only if an IO exception occurs.