Race condition fixed.
We implicitly represented endpoint who were in "initialization state" as an
empty mvar. This does not work, because handleConnectionRequest needs to know
the state of the remote endpoint explicitly (this is the bug we fixed in the
previous commit), but we can't test an mvar for being empty, because this might
equally mean that another thread is in the process of modifying said MVar.
Hence, we introduced an additional constructor RemoteEndPointInit for the
RemoteState ADT, which carries a "resolved MVar (like RemoteEndPointClosing)
which can be used by threads which want to block until the remote endpoint has
been fully initialized.