Write a test that exposes a bug in receiveWait
In particular, a timeout value of 0 means "non-blocking check for matching
messages". Unfortunately, Chan does not support a non-blocking read, so we have
to move to TChan (there is a ticket #4535 for adding tryReadChan, which Simon M
seems to have OKed, but the ticket has subsequently been closed as "Invalid").