More compositional way of supporting 'call'
Rather than statically determine which closures we might potentially 'call',
instead statically generate a mapping from types of the form 'Process a' to
specialized processes (send :: ProcessId -> a -> Process ()) and then do a
runtime lookup when we execute a call.

This might seem like a step backwards (before we didn't need to do a runtime
lookup for which sender to use), but it opens the doors for providing Closure
combinators.