Start tidying up the code
Although we'll still have a central Types file, all other functionality will be
in separate files. This might require some recursive modules -- so be it. Also,
we should make sure that the functions we export from public modules (i.e.,
those not in Internal/) are in fact designed to be in the public API.

This is only the first step towards this goal. To be continued.