Focus on 'Static', not 'Closure'
In particular, provide staticApply as primitive, not closureApply. This
actually makes *much* more sense, as is evidenced by the fact that
resolveClosure is now only a couple of lines, and contains no special treatment
of any function at all; moreover, call and co now have the proper type with

  Static (SerializableDict a) -> ...

and we don't use any trickery to define them; in fact, users can define their
own similar functions just as easily.

This commit represents a signficant rewrite of the entire treatment of
closures, up to the point that the unit tests pass again.

We don't yet provide support for $mkClosure.