TH support: functionTDict. Bugfix in NC.
The node controller maintained its own connection map. This is incorrect: for
once, when process A on node N links to a process B on node M, it sends a
control message to its local node controller, which then forwards the message
to M's node controller. When N forwards the message, it looks up a connection
to N in its own table -- i.e., it's using a connection (N, M). Now suppose M
dies, and restarts, and another process A' links to (possibly) another process
on the new M'. If the local node controller looks up a connection in its own
table, it finds that (N, M) has been blacklisted. However, actually, this is a
message from (A', M) and hence should be using a new connection.