Avoid I/O inside modifyMVar
Many of the API functions had the structure

    modifyMVar (state of heavyweight connection) $ do
      ...
      send on heavyweight connection
      ...
      return (new heavyweight connection state)

Unfortunately, the thread that handles *incoming* messages also occassionally
needs to make modifications to the state of the heavyweight connection, and
hence also calls modifyMVar in various places. This can lead to "distributed
deadlock": consider endpoints A and B. Both A and B start sending lots of
messages to the other party, so much that in both cases the 'send' in the above
example starts to block (because the OS send buffer fills up). This means that
both parties will be stuck, with the MVar "locked", until the other party
receives some messages. However, it might happen that both parties did a
receive and need to obtain the mvar in order to make some changes -- but that
MVar is still blocked by the sending thread on that side. So we have

                   waiting on
  A send thread ---------------\            /------------ B send thread
       ^                        \          /                    ^
       |                         \        /                     |
       | waiting on       /--------------/                      |
       |                 /         \                            |
  A receive thread <----/           \-------------------> B receive thread

A nice distributed deadlock. To avoid this problem we now avoid doing sends
within a modifyMVar. This is not entirely trivial, however. Simply doing

  thingToSend <- modifyMVar (....)
  send thingToSend

is not good enough, because now if we have to threads doing the above the sends
may be done in the wrong order. We therefore do

  act <- modifyMVar (... ; act <- schedule (...); ...; return act)
  runScheduledAction act

'schedule' adds an IO operation to a queue, and runSCheduledAction runs the
next operation from this queue. This may mean of course that
'runScheduledAction' may not execute the action which was scheduled "textually
before" it, so we have to be careful that the right exceptions are thrown in
the right place. That's why 'schedule' creates an MVar for each operation, and
'runScheduledAction' waits for that MVar to be filled.