Optimize handling of 'Crossed' connection requests
When A tries to connect (set up a heavyweight connection to) B, while B tries
to connect to A at the same time, there will be two threads trying to set up a
heavyweight connection for A (and similarly for B): the thread that deals with
the apiConnect, and the thread that deals with the incoming request from B. In
the case that "B" wins and sends a ConnectionRequestCrossed message to A, the
thread that daels with the apiConenct should simply wait until the
handleIncomingRequest thread completes.

We did this already, but it is important that the handleIncomingRequest thread
*in B* (which loses to B's apiConnect thread) waits until A receives the
ConnectionRequestCrossed message (to avoid certain ordering problems). We had
implemented this by regarding an incoming ConnectionRequestCrossed message as
changing the heavyweight connection from Init to Closed state (and signalling
on resolved), but that's wrong as it will cause A to try and connect again. Now
instead when we receive a ConnectionRequestCrossed message we signal on a (new)
MVar "crossed" (part of the Init state), but leave the EndPoint in Init state
(it will be resolved by the handleIncmoingMessage thread).