Switch to asynchronous connect
In the original protocol we had recipient-allocated connection IDs. Hence, when
doing a connect, we would send a connection request to the remote endpoint
(provided that a heavyweight connection is already in place) and then wait for
the recipient to reply with a new connection ID. This however makes connection
much slower than it needs to be. We now use sender-allocated connection IDs, so
that connect turns into a simple message to the remote endpoint that we
allocated a new connection, and we no longer have to wait for a response.

This means that all operations, with the sole exception of creating the
heavyweight connection itself, are now synchronous. Hence, we have completely
removed infrastructure relating to 'remote requests' (doRemoteRequest,
ConnectionRequestResponse, etc.).

What's not obvious is that we also had to change the 'CloseSocket' message.  A
sends a CloseSocket message to B when A thinks that the heavyweight connection
to B is no longer required. A doesn't immediately close the socket, because it
is possible that B has already sent a new connection request (now a
CreatedNewConnection message) to A, even though A has not yet received it.
Instead, it puts the heavyweight connection in 'Closing' state and waits for B
to confirm or deny the request.

When B receives A's CloseSocket message, it checks if it agrees that the
heavyweight connection can be closed. If not, it simply ignores the request and
A will forget it ever sent a CloseSocket request when it receives a
CreatedNewConnection message from B (that is, it moves the endpoint from
Closing to Valid state). If B does agree, then it will reply with a CloseSocket
message, and then actually close the socket. When A receives a CloseSocket
message having sent a CloseSocket message it closes the socket immediately.

However, now that we have asynchronous connection requests we might have the
following potential ordering of messages:

  A sends CloseSocket request, and puts endpoint in closing state

  B sends CreatedNewConnection, CloseConnection, and CloseSocket requests,
  putting its own representation of the endpoint also in closing state.

  B receives the CloseSocket message from A. Since it's endpoint is in Closing
  state, it immediately closes the socket.

  A receives B CreatedNewConnection message, hence *cancelling* the CloseSocket
  message it sent to be and putting the endpoint back in Valid state -- *even
  though* B did closed the socket.

To avoid this problem, we now send the ID of the last created connection along
with the close socket request. Then when B receives a CloseSocket request, it
can compare this ID with the ID of the last connection it opened. If this does
not match, B knows that there is still a CreatedNewConnection message underway
to A, which will cancel the CloseSocket, and hence it will know to ignore the
CloseSocket request from A.