Important bugfix: reconnecting didn't always work
The following sequence resulted in deadlock:

  A connects to B
  A closes connection
  B connects to A
  B closes connection
  B reconnects to A

The problem was in the network protocol, where A tells B the ID of the last
connection B opened to A. If this is lower than the last connection B
*actually* opened, B knows there are still some messages on the way to A and B
must wait before they are resolved. However, this connection ID should be kept
*per heavyweight connection*. We were inconsistent about this: we recorded the
ID of the last created connection on the heavyweight connection, but the next
available ID was a 'global' variable on the local endpoint.

This meant that B sometimes incorrectly concluded that some messages were still
on the way to A, resulting in deadlock.