More support for dealing with errors
However, current setup in the mock network implementation is insufficient,
because it requires the existence of a socket before we can register an error.
That means we can't simulate certain errors (for instance, failure when trying
to create the first connection) because we can't hook into the network layer to
get the socket at that point.