More informative ProcessRegistrationException.
Now includes the identifier of the process that owns the name, if any.
The exception used to only mention the name. However, that made it
impossible to implement many algorithms, such as "return registered
process or spawn new one" in a race free way. If you catch
ProcessRegistrationException after attempting `register`, doing
`whereis` after that to find out who owns the name is racy.

This is a BREAKING CHANGE, albeit to a peripheral part of the current
API, so the next release of distributed-process that includes this
change will have to bump the major version number according to the PVP.