Introduce STM based dequeue method in CQueue
In order to fix all problems that can appear from exception
and timeout handing fully STM based implementation is introduced.

This implementation uses 2 STM transactions, one in order to
read all values from the incomming channel and second in order
to find matched message. This is done because otherwise any
new value in incomming queue can lead to trasanction restart.
And if exception will be rised either in any of 2 transactions
or between then CQueue state will still be consistent.