Workaround race in callLocal.
Since processes spawned with spawnLocal do not inherit the masking state
of the caller, it is possible for the child to be killed before filling
the mvar from which the caller wants to grab the result.

With this patch, when the parent is interrupted, it waits for the child
to die.