Protect callLocal from asynchronous exceptions during the clean up.
Interrupting the cleanup could cause the original exception to be
ignored (i.e. the exception which caused the clean up to be executed).

Ignoring the exception could be problematic. For instance, the exception
can be meant to kill the thread. A later exception comming during the
clean up could come from a wrapping call of System.Timeout.timeout,
which would rather interrupt the computation but certainly not kill the
thread.