Stop using the transport for local communication.
The NC was still using the transport to communicate with itself or with
local processes when using the primitives in C.D.P.I.Messaging.

As a side effect, this patch also stops creating connections involving
processes when the NC sends a message. Only the node-to-node connection
is used by the NC to send messages.