Avoid loop when tracing received messages.
When a trace message was sent to "trace.logger", the NC would produce
a trace message to indicate the arrival of it, this new trace message
would in turn have the NC produce another trace message and so on.

In this patch the NC does not produce a trace message when the
destination is the "trace.logger" process.