Fix for #23 NTTCP-9
The protocol for resolving crossed connections was not able to deal with
connection failures.

This patch implements a probing mechanism which allows the transport to
judge if a connection is healthy and discards it if not.

On the client side, remote endpoints in init state are dropped when
timeouts occur while waiting for the remote side to establish a connection.