Don't use TimeInterval for specifying timeouts.
The lower level primitives to implement timeouts take the amount of
microseconds as an Int. This means there are values of TimeInterval
which would be too large to be converted to this representation.

Users of d-p-async who want to use TimeInterval can still use it to build
the arguments of d-p-async calls.