Have nsend send unencoded messages to local processes.
On one hand, this is an optimization. It saves decoding the message
when the NC handles the message to a local destination. On the other
hand, it guarantees the message will be fully evaluated (up to endoding)
since encoded messages are not fully evaluated unless they go through
the transport.