QDisc is in TCPParameters for whole transport
This was dcoutts's suggestion.
This makes things slightly simpler: you can create your EndPoint using
the typical network-transport API rather than the TCP
TransportInternals.
On the other hand, it means that every EndPoint for the transport will
have the same QDisc, just possibly with different state etc. (this is
why tcpQDisc is in IO). This could make it rather cumbersome to control
queueing parameters for a particular EndPoint, but maybe that's OK and
it's generally better to control the queueing transport-wide?