Define simpleOnePlaceQDisc
This is like simpleUnboundedQDisc except that a single MVar is used
instead of an unbounded Chan. Threads trying to post events will block
on putMVar until some thread clears it using receive on the EndPoint.
It's trusted that the RTS will fairly unblock the putMVars.

One unhappy consequence of this QDisc is that it means connections from
an EndPoint to itself are quite easy to deadlock, if you try to send and
receive in the same thread. Inducing more than one event before
receiving it will deadlock. The test suite therefore does not pass.