Tests receive events eagerly
The motivation for this change: it should not be assumed that events at
an EndPoint will be queued without bound. That's to say, it should not
be assumed that an action which induces an event at an EndPoint will not
block if the previous event has not been pulled out by receive.