apiCloseEndPoint blocks until no longer reciving
When testing components built atop network-transport-tcp, it's sometimes
desirable to have a test case create an end point at the beginning of
the test and close it at the end, where all test cases share the same
transport. If we run such a test repeatedly, as is the typical style
when quick-checking, for instance, then we risk undefined behavior: when
the next test case runs, the end point from the previous test case,
although closed, may still be receiving on one of its sockets. The
sockets have all been closed, their file-descriptors freed up, and so we
get undefined behavior. This can be catastrophic: the stray thread can
steal 'recv's from the next end point's sockets, completely breaking the
protocol and for instance causing the end point to receive an incorrect
length prefix.

When the next end point opens a heavyweight connection, the stray
thread may read the end point identifier (int32), causing the new end
point to read the length prefix of the end point address as the end
point id, and the first 4 bytes of the end point address as the length
of the end point address itself. That will either cause it to get the
wrong end point address, or as I observed, to hang indefinitely because
those first 4 bytes are a very large number.