Configurable limit on bytes in Received event
Since 'QDisc's have been added, it's now possible to control the rate of
incoming events. However, it's not so easy to control the rate of
incoming data, because one 'Received' event could include as many as
2^32 bytes. This patch gives a limit on the number of bytes in any
single 'Received' event. The 'recvExact' function remains, using
'maxBound :: Word32' as the limit, which was the implicit limit anyway.

Also made the maximum data payload and address lengths non-optional; the
default is 'maxBound :: Word32'.