No EndPointAddress equality checks
These have been replaced by a softer kind of equality: connecting to the
EndPointAddress given in a 'ConnectionOpened' or 'EventConnectionLost'
event will indeed reach the peer who opened the connection or to whom
the connection was lost.

This gives network-transport implementations freedom to use
non-bytestring-equal representations. A TCP implementation, for
instance, could include an OS-determined host in the 'EndPointAddress'
given to the events.