Protocol versioning
A candidate implementation of protocol versioning. Here's how it works:

Versionless part of the protocol:

- When `A` connects to `B` it always sends a `Word32` version identifier
  followed by a `Word32` giving the length of remaining data (presumably
  version-specific handshake data).
- `B` can request a different version by sending
  `ConnectionRequestUnsupportedVersion` followed by a `Word32`
  indicating the version which `B` would prefer to use. Since it has the
  length of the remaining data, it can then read and discard the rest of
  the data sent by `A`.
- `A` can try again using `B`'s preferred version, or abandon the
  connection if it doesn't support that version.

If `B` accepts the version number, then the rest of the interaction
proceeds according to that version.

Currently there is only version 0x00000000. Any connection with a
different version number will be rejected as unsupported, and any
rejected connection attempt will abandonned because this version of
nt-tcp supports only 0x00000000.