Support for transports behind NAT
A transport need not have an external address. Random addresses are used
for its endpoints. A random address is generated when such an
unreachable endpoint connects to a reachable endpoint (address is
ephemeral, local to the connected endpoint).

The connection protocol has changed: must send a 0 :: Word32 to
indicate that an address is coming, or 1 :: Word32 to indicate that no
address is coming.