Always handleIncomingMessages on connect
There was a race which caused testSendException
(network-transport-tests) to hang sometimes: if the connection is closed
due to sending an exceptional value, it may put the remote endpoint
state to 'failed' before 'handleIncomingMessages' has begun to receive
on the socket. And so it will nevre receive on the socket, and no
ConnectionLost event will be enqueued.

This is fixed by having handleIncomingMessages take the socket as a
parameter, rather than trying to get it from the shared state.