Fix testReconnect
A delay of 100ms is always taken after the socket closes (closed by the
server), so that sends by the client will fail. If this were not done,
it's possible that the send will go before the OS knows that the server
has closed the TCP connection. This tests would therefore fail
sometimes.