testCloseConnect synchronization
When the server receives on line 663, we expect the connection closed
event from the first client to come through as a result of line 705
(closeEndPoint). But immediately after the closeEndPoint, a new endpoint
was created and a new connection made from that endpoint to the server.
It was possible that the server would receive the connection opened
event from the second endpoint before the connection closed event from
the first endpoint. But now it's not: the second client waits on the
serverAddr MVar which won't be filled until the connection closed event
is dequeued.