Remove test prog, since this patch set is not about console tracing
We're getting intermittent failures in CI (and I've seen some locally) for the tracing suite, but only when run with the TCP transport. I will open a bug to investigate, but since we're not focusing on that in this patch set, I am rolling this change (from master) back and we will test tracing only with the in-memory transport for now.