Fix possible msg corruption on a busy network
The non-atomic sendMany routine from the network pkg can send
one msg via several c_writev calls when the OS sending queues
are busy. So in between these calls the sending thread could
be interrupted by an inner IOException (when, for example,
the connection breaks by the TCP user timeout) or by an outer
exception (like ProcessLinkException). Concurrent threads (which
scheduled the sending before the exception on still valid remote
enpoint) could re-use the socket and start sending a new msg then
which would cause the old msg corruption on the receiving side.

Now we prevent the socket usage at sendOn after an IOException
(with the help of the same MVar which was already there, so we
have it for free). And we call sendMany in a separate thread
which is not targeted (and thus, not interrupted) by the outer
exceptions. (Haskell threads are cheap, aren't they?)